feat: add support for .mjs, .cjs on Node.js 14+
This commit instrucing async + import loading paths to be used on Node.js 14+ runtimes, adding support for .mjs and .cjs files in the process.

Support for the Node.js 10.x runtime is no longer declared on the layer, as it is no longer supported by AWS itself and no new functions can be created with it.